+++ lib/module/once.ts
@@ -9,7 +9,7 @@
 export function once<T extends (...args: any[]) => any>(fn: T): T {
   let isRun = false;
   let result: ReturnType<T>;
-  return function (...args: Parameters<T>): ReturnType<T> {
+  return function (this:any,...args: Parameters<T>): ReturnType<T> {
     if (!isRun) {
       isRun = true;
       result = fn.apply(this, args);