+++ lib/utils/public.ts
@@ -1,3 +1,4 @@
+declare var process: any;
 export const isNode = typeof process !== 'undefined' && process.versions != null && process.versions.node != null;
 export const isBrowser = typeof window !== 'undefined' && typeof document !== 'undefined';